requirements.txt
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
numpy>=1.24.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
tqdm>=4.65.0
requests>=2.31.0